allow "borrowing" keys in Map
