+++ src/containers/mod.rs
@@ -9,13 +9,3 @@ pub trait Storable {
 
     fn access_impl<S>(storage: S) -> Self::AccessorT<S>;
 }
-
-pub trait Key {
-    fn bytes(&self) -> &[u8];
-}
-
-impl Key for str {
-    fn bytes(&self) -> &[u8] {
-        self.as_bytes()
-    }
-}
